proto: split public key backend code by backend
